// src/core/model/interfaces/Produto.ts
export default interface Produto {
  id: string;
  nome: string;
  price: number;
  link: string;
  estoque: boolean;
  descricao: string | null;
  tamanhos: string | null;
  img: string;
  img2: string | null;
  img3: string | null;
  img4: string | null;
  img5: string | null;
  cor1: string | null;
  cor2: string | null;
  cor3: string | null;
  cor4: string | null;
  cor5: string | null;
}